feat: garbage collection feature enable
This change uses the new garbage collection and enables this during the
reconciliation of `AWSCluster` and `AWSManagedControlPlane`. the garbage
collection feature is enabled via a new feature flag
`ExternalResourceGC` which is disabled by default. If the feature flag is
enabled then the the gc service is called in both `reconcileNormal` and
`reconcileDelete` for the infra clusters. The actual gc service does the
work of clearup.

New commands have been added to `clusterawsadm` to allow users to
opt-in/out an already existing cluster from garbage collection.

Additionally, with the new mocks folder introduced with the gc service
the existing mocks have been deleted and tests/controllers updated.

Signed-off-by: Richard Case <richard.case@outlook.com>