refactor: gc opt in/out any time before delete
Changed the proposal based on feedback from review. To quote
@dlipovetsky:

>  In the proposal, the user decides whether to enable the gc when they create the cluster. In the model I describe, the user decides whether to enable the gc at any time before they delete the cluster.

This proposal has been changed so that a user can decide if they want a
workload cluster to opt in or out of garbage collection anytime before
its deleted. The default is that if the gc feature is enabled then a
workload cluster will be deleted.

Signed-off-by: Richard Case <richard.case@outlook.com>